tdf#162289 Update printer info before getting default printer
Move the call to `Printer::updatePrinters` to update the printer
infos from the print dialog ctor to the beginning of
`Printer::PreparePrintJob` (which is the only method
potentially creating a print dialog) to make
sure that the printer infos (including the default printer)
are already up to date there.

This e.g. makes sure that the current CUPS default printer on
Linux gets preselected the next time the print dialog gets
opened if it was changed while LO is running.
(Otherwise, it would only be preselected as expected when
the print dialog gets opened, closed and reopened again.)

For macOS, there was already code to explicitly call
`Printer::updatePrinters` in case the dialog was not shown.
Drop that, as it happens unconditionally at the beginning
now.

Change-Id: I5cd530c0c9801bde88e40f5ac2220feb0e8d4247
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171371
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>