tdf#162282: restore correct handling of EMF
Commit 166a16594eb4e7e78f3f5b00a1432637d109bf8d (Simplify logic in
XOutBitmap::WriteGraphic a bit, 2023-04-01) changed the logic, so
that when the graphic's data reported VectorGraphicDataType::Wmf
type, it wasn't further checked if its link's IsEMF returns true.
That made isKnownVectorFormat return false for such EMFs, and so
the graphic was exported to a new EMF, instead of using its data.
This restores the original behavior.

Change-Id: I68b807e735e501de6bf37172390623938e401e6f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171377
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172063