optimise GetDataEntries a little
instead of swapping variously between std::vector,
std::set<ScTypedStrData> and ScTypedCaseStrSet, just use the last
one, which is the one that we want.

Note that this is a behaviour change, because even though most of
the code carefully used case-sensitive comparison, the code
in ScColumn::GetEntries used the default ScTypedStrData::operator<
which is case-insensitive.

I think that case-sensitive is what we really want.

Change-Id: I097dac6f5b5b1caa1d3dd7000aee516e9923c6e8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171378
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>