Fix signature hash on NSS backend
When using the NSS backend, the SignedAttributes were added after the hash
computation for the timestamp request, altering the final signature and
rendering the timestamp unverifiable.
This patch changes the order in which the attributes are added and uses a copy
of the signer for the hash computation in order to generate a correct hash.

Change-Id: I14cecf1819e80d7a2e470e0f1dee2d09bcbe852c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171384
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit 16a9df36380a24380d4eee77bee3c36223661907)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172816