tdf#162295 DOCX import: optimize the check if the header/footer is empty
The original DOCX bugdoc has some 813 shapes and 150 sections, which
means that isContentEmpty() in the domain mapper is a significant cost.

We don't have layout at import time, so checking if a shape has its
anchor in this header/footer is still expensive, but using internal sw
API makes the 2 other checks constant and the shape check is faster this
way.

- old cost: 16638 ms

- new cost: 11910 ms (72% of baseline)

Change-Id: I22f75ef82b0e2f618e425e076377eea3640b83cc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171390
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins