tdf#162211 writerfilter: don't discard layoutInCell
I DON'T KNOW IF I REALLY WANT TO SUBMIT THIS PATCH.
I'VE ALREADY DECIDED IT IS BEST TO CHANGE THE IMPORT
TO MATCH REALITY, AND KILLING layoutInCell
CERTAINLY DOES THAT.
The Ok button on the UI is the real convincer...
although one unit test had beneficial side effects.

It is very tempting to turn off "layoutInCell".
After all, MSO has rather bizarre exceptions to the rule,
and the whole concept in general is competely unnecessary
(except to fix their horrible initial implementations).

But by discarding layoutInCell,
we are going in the opposite direction
of Microsoft, who in compat15 (2013+),
treats layoutInCell as true
regardless of whether it is set to on or off.

However, there still is one downside to preserving layoutInCell.
In LO's UI, when looking at the properties,
it will change the values to enforce layoutInCell,
so pressing OK will shift an outlying shape.

shape - FRAME/FRAME
-fdo68607.docx [same]
-test_segfault_while_save.docx [zOrder is fixed somehow]
-layout-in-cell-wrapnone-column.docx [same]

make CppunitTest_sw_writerfilter_dmapper \
    CPPUNIT_TEST_NAME=testLayoutInCellWrapnoneColumn

Change-Id: I6d66cb2f14507847e346443d42879a60025bc9d5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171437
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins