tdf#150397 vcl: Don't select SalInstanceTextView text on focus
Disable selecting the text view's text when receiving
keyboard focus (see `TextWindow::GetFocus`) for the
`SalInstanceTextView` implementation.

This aligns this with the GTK implementation
(`GtkInstanceTextView`) that also doesn't do that.

This e.g. prevents selecting the full license
text in the license dialog when installing a
new extension when the qt6 or gen VCL plugins
are used on Linux.

Change-Id: I2f57848f441387fe03abb6d6163416c830cebc4e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171438
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>