simplify IMark hierarchy (1)
The undoes some of the "reduce god objects" work, but leaves the primary
utility of that work in place.

The hierarchy invented to hide the MarkBase hierarchy is an unnecessary
abstraction, and it results in considerable costs in dynamic_cast'ing in
various places, and makes it much harder to optimise and maintain
type-safety.

In this first stage, I move the bulk of the declarations from
bookmark.hxx to IMark.hxx, and remove the IBookmark interface class in
favour of using the concrete Bookmark class (which already exposes some
dead code).

Change-Id: I9712506bd6f58bdddc5193823888e65d24d79ddf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171445
Tested-by: Jenkins
Reviewed-by: Bjoern Michaelsen <bjoern.michaelsen@libreoffice.org>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>