tdf#162075 sw textbox: map a shape's wordwrap=no to the frame's width type
Regression from commit c761df1e42fd11acc5fc05b0baacd803c3788ca6
(tdf#84678 DOCX import: fix handling of textbox margins, 2016-10-25),
the header hape was supposed to have a single line, but it had a
separate line for each character in the shape text.

What happened is that the nominal size of the shape was not enough to
provide enough space for the text, now that we handle spacing between
the frame border and text, and the shape had automatic height set,
instead of automatic width.

Fix the problem by extending SwTextBoxHelper::create() to map the
shape's TextWordWrap property to the inner frame's WidthType. As the
tdf#81567 testcase points out, this requires extending the matching
shape update code as well, because word wrap is always disabled
initially, and it's enabled only after the shape's text is processed.

Also adjust sw/qa/extras/layout/data/tdf138951.odt, which expected word
wrap but didn't request it -- the original tdf#138951 bugdoc keeps
working after this commit.

Change-Id: I17c896f83731a6653e3a0a1493949e27d5259232
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171475
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172109