vcl: Drop obsolete CUPS version check
This CUPS >= 1.7 version check was was introduced back in 2017 in

    commit 7fbf98d99737ac916cc2dc374f3139e2ea81e47b
    Date:   Thu Jul 27 08:11:34 2017 +0530

        Add Common Printing Dialog Functionality

and is obsolete by now.
According to [1], our AlmaLinux baseline has CUPS 2.2.6
by now.

Replace by a simple check whether CUPS is enabled for
the build.
(`pwg_media_t` used in that conditional code block
is from a CUPS header and e.g. not available in
the Android CI builds, see [2].)

[1] https://pkgs.org/search/?q=cups
[2] https://ci.libreoffice.org/job/gerrit_android_x86_64/40875/console

Change-Id: I4683c86c2c9e44ff41ee8093d8f045bc955cba99
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171485
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins