tdf#162340: handle target column scale in setObjectWithInfo
In commit f1121362f3dd4287577dd7952c6f00a524a81111 (tdf#156530: fix
OPreparedStatement::setString, 2024-07-03), I misused the method's
'scale' argument, as the number of decimals in the target column.
It seems that the method was modelled after JDBC's setObject taking
four arguments [1], which use is also unclear to me, but it likely
is intended to describe input data. Anyway, all our implementations
either pass the scale to next implementation, or ignore it.

The problem was when 'x' was a string like '0.1370', and 'scale' was
0. The passed 'scale' was used, instead of the target column's; and
that incorrectly truncated the number to 0.

[1] https://docs.oracle.com/javase/8/docs/api/java/sql/PreparedStatement.html#setObject-int-java.lang.Object-int-int-

Change-Id: Ib507d32706551517cc0a8a3cba351093549d7026
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171506
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>