tdf#148526 sdext,pdfimport: Write clipToStrokePath
'clipToStrokePath' is a variant of 'clip' it sets up a clip
to a path that's been stroked with whatever width the current
pen is.  Now that we have all the rest of the code in, we can
start using it.

This fixes the white blobs on page 3 of tdf#148526 which
are clipped radial fills.  It has a separate problem with
text corruption which this doesn't fix.

It also fixes the geometry of the top left square in:
   https://gitlab.freedesktop.org/poppler/poppler/-/issues/178
(although it still has colour problems with that test case)

Change-Id: Ibe2c56927b45d44e90cfa2934fc905034a50e9c7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172927
Reviewed-by: David Gilbert <freedesktop@treblig.org>
Tested-by: Jenkins