Related: tdf#162295 DOCX import: optimize copying headers/footers
The original DOCX bugdoc has 150 sections on 88 pages, which means that
copyHeaderFooter() in the domain mapper is a significant cost.

Setting a property on a page style involves making a copy of the
SwPageDesc, working on that copy and then copying it back. Which means
that setting UNO properties one by one vs working via
setPropertyValues() is a big difference.

- old cost: 11984 ms

- new cost: 7764 ms (65% of baseline)

Change-Id: Iea99fb3ad9d3e61494c9f012c4c7fa10cc7ab09d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171519
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172050