tdf#160786 sc: Fix repeat char format code inaccuracy
Number format codes allow arbitrary repeat characters. These repeat
characters are inserted while formatting numbers to pad strings to the
width of the containing cell. For example, this is used in some default
format codes to left-align currency symbols with right-aligned digits.

Previously, Calc determined the number of repeat characters to insert
based on the assumption that the pixel width of a string containing N of
the repeat character would be N times the pixel width of a string
containing 1 of them. This is not a safe assumption.

This change updates the algorithm to instead look at the average width
of a repeat character when used in context as part of a padding string.
This approach is still not entirely correct - it doesn't account for
ligatures around the padding sequence, for example - but in practice it
is a reasonable compromise between cost and accuracy.

Change-Id: I240cf38ec0376ee2d2896e7a7da6760a374b3034
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171548
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>