tdf#162351 tdf#130857 qt weld: Implement weld_widget_for_response
... for `QtInstanceMessageDialog`.

This fixes the crash reported in tdf#162351 due to a nullptr
deref in `OSQLMessageBox::impl_addDetailsButton` that was seen
with the base class implementation
`QtInstanceDialog::weld_widget_for_response` that
just returns nullptr.

Clicking the "More" button in that dialog doesn't
yet show the corresponding dialog, but that will
be implemenated separately.

Change-Id: I8cd4fac4fd57be993dba1713bb2e19860dfc334b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171551
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>