tdf#162351 tdf#130857 qt weld: Handle button click
In `QtInstanceButton`, connect the
`QPushButton::clicked` signal with a newly introduced
slot that calls `signal_clicked` from the `weld::Button`
base class, which calls the click handler.

With this in place, clicking the "More" button
in the dialog shown when following the steps from
tdf#162351 now actually results in an "Error Details"
dialog with more information to show up.

Change-Id: I6df43af3ad3bf35b94e712aaaf0158bd7ab57907
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171552
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>