Resolves: tdf#112932 Pressing enter in read-ony Table of Content
doesnt jump to heading

Allows jump to heading using the Enter key or key modifier one in
combination with the Enter key (Ctrl+Enter) when table of content is in
protected mode. This changes the previous behavior of only allowing the
Enter key to jump to the heading when the document is in read-only mode
to also allowing key modifier one in combination with the Enter key to
jump to the heading.

Change-Id: Ib7451cef40d3ecf56295c2b23a6e663fad18fece
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171554
Reviewed-by: Heiko Tietze <heiko.tietze@documentfoundation.org>
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>