sw a11y: Add a11y names for bg settings in side bar
Set a11y names for UI elements related to the page
background in Writer's "Page" sidebar deck, section
"Styles."

Some of those combo boxes/menu buttons are reused for
multiple background types, so set an a11y name
depending on the currently selected background
type in C++ code instead of a static one in the .ui
file for these.

This allows screen readers to announce what those
comboboxes/menu buttons are for when they receive
focus.

Lack of some of those led to gla11y failures in pending
Gerrit change [1], but is actually a preexisting issue
as mentioned in [2].

(The gla11y issue even showed up after simply resaving
pagestylespanel.ui with glade 3.40, see CI failure for [3],
a previous change doing that in preparation for this
commit, but now squashed into this commit.)

[1] https://gerrit.libreoffice.org/c/core/+/171430
[2] https://gerrit.libreoffice.org/c/core/+/171430/comments/6fac1506_50ddcce5
[3] https://gerrit.libreoffice.org/c/core/+/171582

Change-Id: I4e78c61cfd5d5b9f198887eec53be368cb6a0716
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171583
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins