tdf#162121 sw: fix tab stop position in columned ToX
There is a SvxTabAdjust::End but this only exists in the definition of
index entry templates, it is not possible to set this on a tab stop on a
text node.

So the ToxTabStopTokenHandler converts this into a SvxTabAdjust::Right
tab stop with a fixed position.

DefaultToxTabStopTokenHandler::CanUseLayoutRectangle() has inverted
condition but that doesn't matter because there are no layout frames at
that point.
(regression from commit 3aca57fb9c7c979d971cae6bb3ad73c6dc412685)

The main problem is that getting the with from the layout rectangle
never works because there's no layout frame, and the alternative only
takes into account the page dimensions and not any margins of the
section, columns, or gap between columns, not to mention that the ToX
could also be in a table etc.

Refactor this so that the tab stops are set only after all the layout
frames are created.

An unfixable problem remains in case the ToX has columns of different
width.

The wrong tab positions are problematic if TabOverMargin is enabled
since commit 10d753b8aadb50ec4309551b97d4cf2163ea3e3d

Change-Id: Ia712c9bf42b2518e396f1b9e7efd65869ebc5ab4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171585
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Tested-by: Jenkins