tdf#162211 tdf#160077 tdf#91632 layoutInCell: vert page->margin, *->top
Microsoft's layoutInCell is very buggy for vertical orientation.
So, instead of making our own implementation just as buggy,
instead, alter the document so that the settings match the implementation.
That way if MSO ever decides to re-fix layoutInCell,
we are in a position to easily avoid the mapping
to accommodate a new compat version.

make CppunitTest_sw_ooxmlexport21 \
    CPPUNIT_TEST_NAME=testTdf160077_layoutInCellD

Change-Id: Ic02a9b12226fd510656444ed37be470455bc4370
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171433
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>