tdf#162211 revert "tdf#160077 writerfilter: shape vertRelation is FRAME
... for layoutinCell"

This reverts my 24.8 commit d210667c72ff821b8cb50b386a83ed55d65ae9bf
because I'm finding that PRINT_AREA is a better fit than FRAME.

Plus, add a few bits in addition to the revert.
Restrict killing layoutInCell to shapes
where the vertical offset is based on the paragraph too.
(If it is based on the page,
then treating as FRAME only works correctly
if the shape is anchored to the first paragraph.)
Also, just make clear this only makes sense to do
if layoutInCell is not already false, and if isInTable.

Existing unit tests affected were not interesting:
-tdf151704_thinColumnHeight.docx [two off-paper textboxes]
-tdf92157.docx [sync-separated textboxes, not seen in Word2010]

make CppunitTest_sw_ooxmlexport21 \
    CPPUNIT_TEST_NAME=testTdf160077_layoutInCellC

Change-Id: I25e9a68f2bc096bddec6ba9d0dc1bb13a226999e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171339
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171693
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>