SwContentIndex owner pointers
we already have two things that "own" a SwContentIndex, for which
we store a pointer in the SwContentIndex. I want to add more things,
so rather than add more fields to a very heavily used structure,
generalise the mechanism so we can store a single field (since
only thing at a time will ever own a SwContentIndex).

Also add an enum and a GetOwnerType() method so I can avoid
dynamic_cast, this is a pretty performance-sensitive area.

Change-Id: Ifb03a06b0bb69a325f41a8041e9d23cfc82c3268
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171606
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>