tdf#161909: mention that keys/certs preselection is for ODF only
ODFs can be signed with GPG or X.509, so both key types are listed for
selection as a permanent setting in the options.

PDF signing doesn't respect the preselected key/cert. And that makes
sense, because PDFs can only be signed with X.509. So a preselected
GPG key wouldn't work.

Change-Id: I1b5dc36e40a9a2c82ca9044bf1b481968982befd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171636
Tested-by: Jenkins
Reviewed-by: Sarper Akdemir <sarper.akdemir@allotropia.de>