a11y: Clarify XAccessibleHyperlink doc
The API looks like it was modeled after
IAccessible2's `IAccessibleHyperlink` [1].

The doc was saying

> /** Returns an object that represents the link anchor, as appropriate
>     for that link.

for both, `XAccessibleHyperlink::getAccessibleActionAnchor`
and `XAccessibleHyperlink::getAccessibleActionObject`.

However, `XAccessibleHyperlink::getAccessibleActionObject` can
be used to get the link target, not the anchor, i.e. is the
equivalent of `IAccessibleHyperlink::anchorTarget` while
`XAccessibleHyperlink::getAccessibleActionAnchor` is the
equivalent of ` IAccessibleHyperlink::anchor`.

[1] https://accessibility.linuxfoundation.org/a11yspecs/ia2/docs/html/interface_i_accessible_hyperlink.html

Change-Id: I7e28ac648f698592afb031fcb433f385f902c6a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171635
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins