sw: HTML import: apply existing hard formatting at insert position
When selecting a placeholder field and pasting some HTML over it, the
existing formatting of the placeholder field is inconsistently retained
on the first paragraph of the pasted text, but not on subsequent ones.

Fix this by retaining it on all of the pasted text; copy it after doing
the 2 SplitNode() calls in SwHTMLParser::CallParser() and apply it
directly to new paragraphs in SwHTMLParser::SetTextCollAttrs() after
applying the paragraph style (which resets hard formatting).

Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171648
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit f45d2fa85ba89eb7d71b51c3324f5938509b3201)

Change-Id: I8c3596372b415fece46f6bb018cb8c26ff89c5ad