tdf#160583 sw HTML export: fix empty field content on copy
Regression from e09d7acd6148b8417487a4473fdc038c7e30d1c4 (sw: fix crash
in OutHTML_SwFormatField(), 2023-04-24), copy a mail merge field to the
clipboard using the HTML export, the value of that field is now lost.

Digging deeper, this started with
1916d161902bdd52b8cfa5b29153c8f8c39fce52 (De-static-izing colors in
SwViewOption, 2023-03-16), previously the field shading setting was not
part of the view, so the HTML copy of the document could take it into
account, even if clipboard documents don't have a view.

Fix the problem by shrinking the scope of the "do we have a doc shell"
check, so field content is exported again, we just assume that no
shading is wanted for clipboard documents.

Test this via getTransferData(), normal HTML export would work on the
non-clipboard document that has a doc shell, and then the problem is not
visible.

Change-Id: Ib9094eb1d5fc51819279fc58c442e5a507ef2e69
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171661
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>