ScriptForge (SFDialogs) Closure of non-modal dialogs
Before LibreOffice 24.2 the window
Close button was inoperant.
Now a com.sun.star.awt.XTopWindowListener
listener is set on the dialog closure
by the
   dialog.Execute(Modal := False)
method.

In summary, when the window Close button
is clicked:
- a modal dialog is stopped as if a Cancel
  button was pressed
- a non-modal dialog is made hidden

Termination and disposal of the dialog instance
remain roles for user scripts.

This change requires minor changes in the help
documentation.

Change-Id: Ia4a3e39422b4050fc48f5552fadecfd984e46524
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171703
Reviewed-by: Jean-Pierre Ledure <jp@ledure.be>
Tested-by: Jenkins