tdf#91632 doc import layoutInCell: vert page->margin, *->top
The same thing was done early in this series of patches for DOCX

Microsoft's layoutInCell is very buggy for vertical orientation.
So, instead of making our own implementation just as buggy,
instead, alter the document so that the settings match the implementation.
That way if MSO ever decides to re-fix layoutInCell,
we are in a position to easily avoid the mapping
to accommodate a new compat version.

make CppunitTest_sw_ww8export4 \
    CPPUNIT_TEST_NAME=testTdf91632_layoutInCellD

Change-Id: I1a8ad18872ba7c0f08cc5dff517be6e6079e8b11
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171705
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>