Related: tdf#154959 suitable label printer paper size not auto-selected
a) If the printer is the default, then it doesn't get a paper size auto
selected in the paper combobox, even though the preview shows it rendered
on a paper size that the printer supports.

b) With that fixed, then if the printer is the default one then right
paper gets selected automatically, but starting with another and then
switching to it gets different results, so throw away the "page cache"
on switching pages and get it again.

Change-Id: Idd5254eac2668164b02b41a63943f65a222656d7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171721
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>