tdf#162426: make sure to only output new tags after all attributes were written
In commit 03fe839d6e76ed8b1dfb65093ab59a8904852ff6 (html export: rework image
output to use HTML writer, 2014-09-22), optional export of linebreak tag was
implemented in the end of SwHTMLWriter::writeFrameFormatOptions. But it is
called in the middle of img arrtibutes output in OutHTML_ImageStart; so when
br was output, all following img attributes got discarded.

This makes writing the tag a separate function with appropriate name, which
is called after all attributes are already written.

Change-Id: I2a7df56e73150d93c565a7240c047b09f3336641
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171722
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
Signed-off-by: Xisco Fauli <xiscofauli@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172064