tdf#158556 move m_vIndices to SwNode
When we load complex documents we end up with very large
amounts of SwNodeIndex objects, and we also need call
RemoveNode often, which means we end up scanning tons of objects.

So move the linked list from SwNodes to SwNode, so we can
scan just the objects that we are interested in.

Shaves 10% off the load time of a complex docx file.

Change-Id: Id62388dbc7e41fae30acb7910a982710c80e563e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171888
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>
Tested-by: Jenkins