tdf#130857 VclBuilder: Move size group parsing to base class
Move the parsing of size groups from `VclBuilder` to
the base class, `BuilderBase` and add a getter,
`BuilderBase::getSizeGroups` that currently gets
used from `VclBuilder`.

Since the parsing is not specific to vcl::Window,
make it possible to reuse it from other subclasses.
It could e.g. be useful for the WIP
QtInstanceBuilder from Omkar's WIP change [1].

[1] https://gerrit.libreoffice.org/c/core/+/161831

Change-Id: I0c44bfda1d5c850679695ab6bd436e4e487df9c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171819
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins