tdf#156211 sw: Fix spurious kashida inserted after undo
This change fixes an issue presenting as incorrectly-positioned kashida
glyphs overlapping Arabic text after certain edit operations.

During layout with kashida justification, Writer builds a table of lines
that require fallback to whitespace justification. Normally, this table
is built sequentially from the first line, but it may be updated
out-of-order following certain edit operations.

Due to an off-by-one error, if Writer cleared the exclusion for a line
immediately before a legitimately-excluded line, Writer would also clear
the legitimate exclusion. In such a situation, portions of excluded text
would be redrawn with kashida justification, and because that text
usually does not have enough free space, the kashida glyphs would be
drawn on top of the base text.

Change-Id: I204661286531fa6064f7a6adc35f1606e35e5d39
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171878
Tested-by: Jenkins
Reviewed-by: Jonathan Clark <jonathan@libreoffice.org>