tdf#90733 - support invertIfNegative for bar and bubble chart
Extend the existing initial support for the OOXML invertIfNegative tag to pass
the data down into bar and bubble chart rendering. Also add corresponding unit
tests. This does not include a UI control to select/deselect the invert-if-
negative option, nor support for export of the flag to OOXML, nor any support
in ODF.

Change-Id: I45b24b816edc379c9b431e86269dd5ff37977b89
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171879
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>